pyproject.toml
[project]
name = "beep-boop"
version = "0.1.0"
description = "Conversational AI chat bot with agentic capabilities"
authors = [{ name = "Tibo Cintio", email = "tibocin@pm.me" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-dotenv>=1.0.0",
  "openai>=1.0.0",
  "pyyaml>=6.0.0",
  "numpy>=1.21.0",
  "scikit-learn>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "black>=23.0.0", "isort>=5.0.0", "flake8>=6.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["modules"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
